requirements.txt
pandas
dvc
dagshub
scikit-learn
mlflow
dvc-s3
nltk
tensorflow
pyyaml
psutil

